refactor(config): extract coordinator-resolvers.ts (R3)
Move YamlCoordinator, resolveCoordinatorPlanning, resolveCoordinatorDispatchKind,
resolveCoordinatorDispatch, resolveCoordinatorActivation, resolveToolSelectionStrategy,
FinalizerYaml, and buildFinalizer byte-for-byte into
pipelines/coordinator-resolvers.ts. config.ts now barrel-re-exports all public
symbols; pipelines/parsers.ts imports direct from ./coordinator-resolvers.js.
Adds build-finalizer.test.ts (5 cases, incl. no-LLM-config error branch).
